+++ src/main/java/com/appdynamics/extensions/logmonitor/LogMonitorTask.java
@@ -11,25 +11,25 @@ package com.appdynamics.extensions.logmonitor;
 import com.appdynamics.extensions.AMonitorTaskRunnable;
 import com.appdynamics.extensions.MetricWriteHelper;
 import com.appdynamics.extensions.conf.MonitorContextConfiguration;
+import com.appdynamics.extensions.logging.ExtensionsLoggerFactory;
 import com.appdynamics.extensions.logmonitor.config.Log;
+import com.appdynamics.extensions.logmonitor.metrics.LogMetrics;
 import com.appdynamics.extensions.logmonitor.processors.FilePointerProcessor;
 import com.appdynamics.extensions.logmonitor.processors.LogFileManager;
-import com.appdynamics.extensions.metrics.Metric;
 import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import java.util.Map;
+import java.util.List;
 
-import static com.appdynamics.extensions.logmonitor.LogMonitor.metrics;
+import static com.appdynamics.extensions.logmonitor.util.Constants.SCHEMA_NAME;
 import static com.appdynamics.extensions.logmonitor.util.LogMonitorUtil.getFinalMetricList;
-import static com.appdynamics.extensions.logmonitor.util.LogMonitorUtil.resetRegisteredMetricOccurrences;
+import static com.appdynamics.extensions.logmonitor.util.LogMonitorUtil.prepareEventsForPublishing;
 
 /**
  * @author Aditya Jagtiani
  */
 
 public class LogMonitorTask implements AMonitorTaskRunnable {
-    private static Logger LOGGER = LoggerFactory.getLogger(LogMonitorTask.class);
+    private static Logger LOGGER = ExtensionsLoggerFactory.getLogger(LogMonitorTask.class);
     private MetricWriteHelper metricWriteHelper;
     private MonitorContextConfiguration monitorContextConfiguration;
     private Log log;
@@ -53,13 +53,25 @@ public class LogMonitorTask implements AMonitorTaskRunnable {
 
     public void onTaskComplete() {
         LOGGER.info("Completed the Log Monitoring task for log : " + log.getDisplayName());
-        resetRegisteredMetricOccurrences(metrics);
     }
 
     private void populateAndPrintMetrics() throws Exception {
         LogFileManager logFileManager = new LogFileManager(filePointerProcessor, log, monitorContextConfiguration);
-        logFileManager.processLogMetrics();
-        metricWriteHelper.transformAndPrintMetrics(getFinalMetricList(metrics));
+        LogMetrics logMetrics = logFileManager.processLogMetrics();
+        publishEvents(logMetrics);
+        LOGGER.info("Printing {} metrics for Log {}", logMetrics.getMetrics().size(), log.getDisplayName());
+        metricWriteHelper.transformAndPrintMetrics(getFinalMetricList(logMetrics.getMetrics()));
         filePointerProcessor.updateFilePointerFile();
+    }
+
+    private void publishEvents(LogMetrics logMetrics) {
+        List<LogEvent> events = logMetrics.getEventsToBePublished();
+        if(events.size() != 0) {
+            List<String> eventsToBePublished = prepareEventsForPublishing(events);
+            monitorContextConfiguration.getContext().getEventsServiceDataManager().publishEvents(SCHEMA_NAME, eventsToBePublished);
+        }
+        else {
+            LOGGER.info("No events to publish for log {}, skipping", log.getDisplayName());
+        }
     }
 }