+++ src/main/java/com/appdynamics/extensions/logmonitor/processors/LogFileManager.java
@@ -8,8 +8,10 @@
 
 package com.appdynamics.extensions.logmonitor.processors;
 
-import com.appdynamics.extensions.MonitorExecutorService;
 import com.appdynamics.extensions.conf.MonitorContextConfiguration;
+import com.appdynamics.extensions.eventsservice.EventsServiceDataManager;
+import com.appdynamics.extensions.executorservice.MonitorExecutorService;
+import com.appdynamics.extensions.logging.ExtensionsLoggerFactory;
 import com.appdynamics.extensions.logmonitor.config.FilePointer;
 import com.appdynamics.extensions.logmonitor.config.Log;
 import com.appdynamics.extensions.logmonitor.metrics.LogMetrics;
@@ -19,7 +21,6 @@ import org.apache.commons.io.filefilter.WildcardFileFilter;
 import org.apache.commons.lang.StringUtils;
 import org.bitbucket.kienerj.OptimizedRandomAccessFile;
 import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.io.File;
 import java.io.FileFilter;
@@ -38,11 +39,13 @@ import static com.appdynamics.extensions.logmonitor.util.LogMonitorUtil.*;
  */
 
 public class LogFileManager {
-    private static final Logger LOGGER = LoggerFactory.getLogger(LogFileManager.class);
+    private static final Logger LOGGER = ExtensionsLoggerFactory.getLogger(LogFileManager.class);
     private Log log;
     private FilePointerProcessor filePointerProcessor;
     private MonitorContextConfiguration monitorContextConfiguration;
     private MonitorExecutorService executorService;
+    private EventsServiceDataManager eventsServiceDataManager;
+    private int offset;
 
     public LogFileManager(FilePointerProcessor filePointerProcessor, Log log,
                           MonitorContextConfiguration monitorContextConfiguration) {
@@ -65,7 +68,8 @@ public class LogFileManager {
                 String dynamicLogPath = dirPath + log.getLogName();
                 long currentTimeStampFromFilePointer = getCurrentTimeStampFromFilePointer(dynamicLogPath, file.getPath());
                 long currentFilePointerPosition = getCurrentFilePointerOffset(dynamicLogPath, file.getPath());
-
+                eventsServiceDataManager = evaluateEventsServiceConfig();
+                offset = (Integer) this.monitorContextConfiguration.getConfigYml().get("logMatchOffset");
                 if (hasLogRolledOver(dynamicLogPath, file.getPath(), file.length())) {
                     List<File> filesToBeProcessed = getFilesToBeProcessedFromDirectory(currentTimeStampFromFilePointer, dirPath);
                     latch = new CountDownLatch(filesToBeProcessed.size());
@@ -97,7 +101,7 @@ public class LogFileManager {
                 randomAccessFile.seek(0);
             }
             executorService.execute("LogMetricsProcessor", new LogMetricsProcessor(randomAccessFile, log, latch,
-                    logMetrics, currentFile, getMetricCharacterReplacers()));
+                    logMetrics, currentFile, eventsServiceDataManager, offset));
         }
     }
 
@@ -109,14 +113,15 @@ public class LogFileManager {
         OptimizedRandomAccessFile randomAccessFile = new OptimizedRandomAccessFile(file, "r");
         randomAccessFile.seek(currentFilePointerPosition);
         executorService.execute("LogMetricsProcessor", new LogMetricsProcessor(randomAccessFile, log, latch, logMetrics,
-                file, getMetricCharacterReplacers()));
+                file, eventsServiceDataManager, offset));
     }
 
     private void setNewFilePointer(String dynamicLogPath, CopyOnWriteArrayList<FilePointer> filePointers) {
         FilePointer latestFilePointer = LogMonitorUtil.getLatestFilePointer(filePointers);
         LOGGER.debug("Updating File Pointer with the most recently processed log: {}, pointing to file: {} with the " +
                         "last read position: {} and a creation time stamp of: {}", dynamicLogPath,
-                latestFilePointer.getFilename(), latestFilePointer.getLastReadPosition(), latestFilePointer.getFileCreationTime());
+                latestFilePointer.getFilename(), latestFilePointer.getLastReadPosition(),
+                latestFilePointer.getFileCreationTime());
         filePointerProcessor.updateFilePointer(dynamicLogPath, latestFilePointer.getFilename(),
                 latestFilePointer.getLastReadPosition(), latestFilePointer.getFileCreationTime());
     }
@@ -126,12 +131,12 @@ public class LogFileManager {
         if (!resolvedPath.endsWith(File.separator)) {
             resolvedPath = resolvedPath + File.separator;
         }
+
         return resolvedPath;
     }
 
     private long getCurrentTimeStampFromFilePointer(String dynamicLogPath, String actualLogPath) {
-        FilePointer filePointer =
-                filePointerProcessor.getFilePointer(dynamicLogPath, actualLogPath);
+        FilePointer filePointer = filePointerProcessor.getFilePointer(dynamicLogPath, actualLogPath);
         return filePointer.getFileCreationTime();
     }
 
@@ -217,14 +222,15 @@ public class LogFileManager {
         return filePointerProcessor.getFilePointer(dynamicLogPath, actualLogPath).getLastReadPosition().get();
     }
 
-    private Map<Pattern, String> getMetricCharacterReplacers() {
-        List<Map<String, String>> metricCharReplacersFromCfg = (List) monitorContextConfiguration.getConfigYml()
-                .get("metricCharacterReplacers");
-        return initializeMetricCharacterReplacers(metricCharReplacersFromCfg);
-    }
-
     private void handleFileEncoding(File file) throws Exception {
         LOGGER.debug("Converting current file: {} to UTF-8 encoding for further processing", file.getName());
         convertToUTF8Encoding(file, log.getEncoding());
+    }
+
+    private EventsServiceDataManager evaluateEventsServiceConfig() {
+        if (this.monitorContextConfiguration.getConfigYml().get("sendDataToEventsService").equals(true)) {
+            return monitorContextConfiguration.getContext().getEventsServiceDataManager();
+        }
+        return null;
     }
 }