+++ src/test/java/com/appdynamics/extensions/logmonitor/processors/LogFileManagerTest.java
@@ -10,19 +10,21 @@ package com.appdynamics.extensions.logmonitor.processors;
 
 import com.appdynamics.extensions.AMonitorJob;
 import com.appdynamics.extensions.conf.MonitorContextConfiguration;
-import com.appdynamics.extensions.logmonitor.LogMonitor;
+import com.appdynamics.extensions.eventsservice.EventsServiceDataManager;
+import com.appdynamics.extensions.logmonitor.LogEvent;
 import com.appdynamics.extensions.logmonitor.config.FilePointer;
 import com.appdynamics.extensions.logmonitor.config.Log;
+import com.appdynamics.extensions.logmonitor.config.SearchPattern;
 import com.appdynamics.extensions.logmonitor.config.SearchString;
 import com.appdynamics.extensions.logmonitor.metrics.LogMetrics;
 import com.appdynamics.extensions.logmonitor.util.LogMonitorUtil;
 import com.appdynamics.extensions.metrics.Metric;
 import com.google.common.collect.Lists;
 import org.apache.commons.lang.StringUtils;
-import org.junit.Before;
+import org.bitbucket.kienerj.OptimizedRandomAccessFile;
+import org.junit.Assert;
 import org.junit.Test;
 import org.mockito.Mockito;
-import org.powermock.reflect.Whitebox;
 
 import java.io.*;
 import java.nio.channels.FileChannel;
@@ -30,9 +32,10 @@ import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
+import java.util.regex.Pattern;
 
-import static org.junit.Assert.assertEquals;
+import static com.appdynamics.extensions.logmonitor.util.Constants.SCHEMA_NAME;
+import static org.junit.Assert.*;
 import static org.mockito.Matchers.anyString;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.when;
@@ -46,15 +49,7 @@ public class LogFileManagerTest {
     private FilePointerProcessor mockFilePointerProcessor = Mockito.mock(FilePointerProcessor.class);
     private MonitorContextConfiguration monitorContextConfiguration = new MonitorContextConfiguration("Log Monitor",
             "Custom Metrics|Log Monitor|", Mockito.mock(File.class), Mockito.mock(AMonitorJob.class));
-    private Map<String, Metric> metrics;
 
-    @Before
-    public void setUp() {
-        metrics = new ConcurrentHashMap<String, Metric>();
-        Whitebox.setInternalState(LogMonitor.class, metrics);
-    }
-
-    //region <Print Matched String Flag Tests>
     @Test
     public void testProcessorWhenPrintMatchedStringIsFalse() throws Exception {
         Log log = new Log();
@@ -94,7 +89,8 @@ public class LogFileManagerTest {
         monitorContextConfiguration.setConfigYml("src/test/resources/conf/config.yaml");
 
         classUnderTest = new LogFileManager(mockFilePointerProcessor, log, monitorContextConfiguration);
-        classUnderTest.processLogMetrics();
+        LogMetrics logMetrics = classUnderTest.processLogMetrics();
+        Map<String, Metric> metrics = logMetrics.getMetrics();
 
         assertEquals("13", metrics.get("TestLog|Search String|Debug|Occurrences").getMetricValue());
         assertEquals("24", metrics.get("TestLog|Search String|Info|Occurrences").getMetricValue());
@@ -138,7 +134,9 @@ public class LogFileManagerTest {
         filePointer.setFilename(log.getLogDirectory() + log.getLogName());
         when(mockFilePointerProcessor.getFilePointer(anyString(), anyString())).thenReturn(filePointer);
         classUnderTest = new LogFileManager(mockFilePointerProcessor, log, monitorContextConfiguration);
-        classUnderTest.processLogMetrics();
+        classUnderTest = new LogFileManager(mockFilePointerProcessor, log, monitorContextConfiguration);
+        LogMetrics logMetrics = classUnderTest.processLogMetrics();
+        Map<String, Metric> metrics = logMetrics.getMetrics();
 
         assertEquals("13", metrics.get("TestLog|Search String|Debug|Occurrences").getMetricValue());
         assertEquals("24", metrics.get("TestLog|Search String|Info|Occurrences").getMetricValue());
@@ -192,17 +190,15 @@ public class LogFileManagerTest {
         monitorContextConfiguration.setConfigYml("src/test/resources/conf/config.yaml");
 
         classUnderTest = new LogFileManager(mockFilePointerProcessor, log, monitorContextConfiguration);
-        classUnderTest.processLogMetrics();
-
+        LogMetrics logMetrics = classUnderTest.processLogMetrics();
+        Map<String, Metric> metrics = logMetrics.getMetrics();
         assertEquals("5", metrics.get("TestUTF16Log|Search String|Debug|Occurrences").getMetricValue());
         assertEquals("5", metrics.get("TestUTF16Log|Search String|Info|Occurrences").getMetricValue());
         assertEquals("0", metrics.get("TestUTF16Log|Search String|Error|Occurrences").getMetricValue());
 
         revertToUTF16Encoding(new File(log.getLogDirectory() + log.getLogName()));
     }
-    //endregion
 
-    //region <Regex Tests>
     @Test
     public void testProcessorForRegexPatternMatch() throws Exception {
         Log log = new Log();
@@ -260,7 +256,8 @@ public class LogFileManagerTest {
         when(mockFilePointerProcessor.getFilePointer(anyString(), anyString())).thenReturn(filePointer);
 
         classUnderTest = new LogFileManager(mockFilePointerProcessor, log, monitorContextConfiguration);
-        classUnderTest.processLogMetrics();
+        LogMetrics logMetrics = classUnderTest.processLogMetrics();
+        Map<String, Metric> metrics = logMetrics.getMetrics();
 
         assertEquals("5", metrics.get("TestLog|Search String|Pattern <|Matches|<").getMetricValue());
         assertEquals("6", metrics.get("TestLog|Search String|Pattern >|Matches|>").getMetricValue());
@@ -316,7 +313,8 @@ public class LogFileManagerTest {
         when(mockFilePointerProcessor.getFilePointer(anyString(), anyString())).thenReturn(filePointer);
 
         classUnderTest = new LogFileManager(mockFilePointerProcessor, log, monitorContextConfiguration);
-        LogMetrics result = classUnderTest.processLogMetrics();
+        LogMetrics logMetrics = classUnderTest.processLogMetrics();
+        Map<String, Metric> metrics = logMetrics.getMetrics();
 
         // matches (\\s|^)m\\w+(\\s|$)
         assertEquals("7", metrics.get("TestLog|Search String|Pattern start with M|Matches|Memorymetricgenerator").getMetricValue());
@@ -338,9 +336,9 @@ public class LogFileManagerTest {
         assertEquals(getFileSize(log.getLogDirectory(), log.getLogName()),
                 metrics.get("TestLog|File size (Bytes)").getMetricValue());
     }
-    //endregion
+    
 
-    //region <Log Manipulation and Rollover Tests>
+    
     @Test
     public void testLogFileAfterAdditionOfMoreLogs() throws Exception {
         String originalFilePath = this.getClass().getClassLoader().getResource("test-log-1.log").getPath();
@@ -384,6 +382,7 @@ public class LogFileManagerTest {
 
         classUnderTest = new LogFileManager(mockFilePointerProcessor, log, monitorContextConfiguration);
         LogMetrics result = classUnderTest.processLogMetrics();
+        Map<String, Metric> metrics = result.getMetrics();
 
         assertEquals("13", metrics.get("TestLog|Search String|Debug|Matches|Debug").getMetricValue());
         assertEquals("24", metrics.get("TestLog|Search String|Info|Matches|Info").getMetricValue());
@@ -413,9 +412,9 @@ public class LogFileManagerTest {
                 new Date() + "	DEBUG	This is the fifth line");
         updateLogFile(testFilepath, logsToAdd);
 
-        setUp();
         classUnderTest = new LogFileManager(mockFilePointerProcessor, log, monitorContextConfiguration);
         result = classUnderTest.processLogMetrics();
+        metrics = result.getMetrics();
 
         assertEquals("3", metrics.get("TestLog|Search String|Debug|Matches|Debug").getMetricValue());
         assertEquals("2", metrics.get("TestLog|Search String|Info|Matches|Info").getMetricValue());
@@ -429,84 +428,8 @@ public class LogFileManagerTest {
     }
 
     @Test
-    public void testLogFileProcessingAfterRollover() throws Exception {
-        String dynamicLog1 = this.getClass().getClassLoader().getResource("dynamic-log-1.log").getPath();
-
-        String testFilename = "active-dynamic-log-1.log";
-        String testFilepath = String.format("%s%s%s", getTargetDir().getPath(), File.separator, testFilename);
-        copyFile(dynamicLog1, testFilepath);
-
-        Log log = new Log();
-        log.setLogDirectory(getTargetDir().getPath());
-        log.setLogName("active-dynamic-*");
-
-        SearchString searchString = new SearchString();
-        searchString.setCaseSensitive(false);
-        searchString.setMatchExactString(true);
-        searchString.setPattern("debug");
-        searchString.setDisplayName("Debug");
-        searchString.setPrintMatchedString(false);
-
-        SearchString searchString1 = new SearchString();
-        searchString1.setCaseSensitive(false);
-        searchString1.setMatchExactString(true);
-        searchString1.setPattern("error");
-        searchString1.setDisplayName("Error");
-        searchString1.setPrintMatchedString(false);
-
-        log.setSearchStrings(Lists.newArrayList(searchString, searchString1));
-
-        FilePointer filePointer = new FilePointer();
-        filePointer.setFilename(log.getLogDirectory() + File.separator + testFilename);
-        monitorContextConfiguration.setConfigYml("src/test/resources/conf/config.yaml");
-        when(mockFilePointerProcessor.getFilePointer(anyString(), anyString())).thenReturn(filePointer);
-
-        classUnderTest = new LogFileManager(mockFilePointerProcessor, log, monitorContextConfiguration);
-        classUnderTest.processLogMetrics();
-
-        assertEquals("3", metrics.get("active-dynamic-*|Search String|Debug|Occurrences").getMetricValue());
-        assertEquals("0", metrics.get("active-dynamic-*|Search String|Error|Occurrences").getMetricValue());
-
-        String filesize = getFileSize(log.getLogDirectory(), testFilename);
-        assertEquals(filesize, metrics.get("active-dynamic-*|File size (Bytes)").getMetricValue());
-
-        // simulate a file pointer update
-        filePointer.updateLastReadPosition(new Long(filesize));
-        when(mockFilePointerProcessor.getFilePointer(anyString(), anyString()))
-                .thenReturn(filePointer);
-
-        List<String> logsToAdd = Lists.newArrayList();
-        for (int i = 0; i < 100; i++) {
-            logsToAdd.add(new Date() + "	DEBUG	Statement " + i + "\n");
-        }
-        updateLogFile(testFilepath, logsToAdd);
-
-        // simulate new file created with different name
-        Thread.sleep(1000);
-        String dynamicLog2 = this.getClass().getClassLoader().getResource("dynamic-log-2.log").getPath();
-
-        testFilename = "active-dynamic-log-2.log";
-        testFilepath = String.format("%s%s%s", getTargetDir().getPath(), File.separator, testFilename);
-        copyFile(dynamicLog2, testFilepath);
-
-        // simulate another file created with different name
-        Thread.sleep(1000);
-        String dynamicLog3 = this.getClass().getClassLoader().getResource("dynamic-log-3.log").getPath();
-        testFilename = "active-dynamic-log-3.log";
-        testFilepath = String.format("%s%s%s", getTargetDir().getPath(), File.separator, testFilename);
-        copyFile(dynamicLog3, testFilepath);
-        logsToAdd.clear();
-
-        setUp();
-        classUnderTest.processLogMetrics();
-
-        assertEquals("7", metrics.get("active-dynamic-*|Search String|Error|Occurrences").getMetricValue());
-        assertEquals("103", metrics.get("active-dynamic-*|Search String|Debug|Occurrences").getMetricValue());
-    }
-
-    @Test
     public void testFilePointerHasLatestTimeStampAfterRollover() throws Exception {
-        String dynamicLog1 = this.getClass().getClassLoader().getResource("dynamic-log-1.log").getPath();
+        String dynamicLog1 = "src/test/resources/dynamic-log-1.log";
 
         String testFilename = "active-dynamic-log-1.log";
         String testFilepath = String.format("%s%s%s", getTargetDir().getPath(), File.separator, testFilename);
@@ -538,10 +461,10 @@ public class LogFileManagerTest {
         when(mockFilePointerProcessor.getFilePointer(anyString(), anyString())).thenReturn(filePointer);
 
         classUnderTest = new LogFileManager(mockFilePointerProcessor, log, monitorContextConfiguration);
-        LogMetrics result = classUnderTest.processLogMetrics();
+        LogMetrics logMetrics = classUnderTest.processLogMetrics();
 
         String filesize = getFileSize(log.getLogDirectory(), testFilename);
-        FilePointer latestFilePointer = LogMonitorUtil.getLatestFilePointer(result.getFilePointers());
+        FilePointer latestFilePointer = LogMonitorUtil.getLatestFilePointer(logMetrics.getFilePointers());
         Mockito.verify(mockFilePointerProcessor, times(1))
                 .updateFilePointer("./target/active-dynamic-*",
                         latestFilePointer.getFilename(), latestFilePointer.getLastReadPosition(), latestFilePointer.getFileCreationTime());
@@ -576,14 +499,146 @@ public class LogFileManagerTest {
         }
 
         updateLogFile(testFilepath, logsToAdd);
-        result = classUnderTest.processLogMetrics();
-        latestFilePointer = LogMonitorUtil.getLatestFilePointer(result.getFilePointers());
+        logMetrics = classUnderTest.processLogMetrics();
+        latestFilePointer = LogMonitorUtil.getLatestFilePointer(logMetrics.getFilePointers());
         Mockito.verify(mockFilePointerProcessor, times(1)).updateFilePointer("./target/active-dynamic-*",
                 latestFilePointer.getFilename(), latestFilePointer.getLastReadPosition(), latestFilePointer.getFileCreationTime());
     }
-    //endregion
 
-    //region <Test Utilities>
+    @Test
+    public void testLogFileProcessingAfterRollover() throws Exception {
+        String dynamicLog1 = "src/test/resources/dynamic-log-1.log";
+        String testFilename = "active-dynamic-log-1.log";
+        String testFilepath = String.format("%s%s%s", getTargetDir().getPath(), File.separator, testFilename);
+        copyFile(dynamicLog1, testFilepath);
+
+        Log log = new Log();
+        log.setLogDirectory(getTargetDir().getPath());
+        log.setLogName("active-dynamic-*");
+
+        SearchString searchString = new SearchString();
+        searchString.setCaseSensitive(false);
+        searchString.setMatchExactString(true);
+        searchString.setPattern("debug");
+        searchString.setDisplayName("Debug");
+        searchString.setPrintMatchedString(false);
+
+        SearchString searchString1 = new SearchString();
+        searchString1.setCaseSensitive(false);
+        searchString1.setMatchExactString(true);
+        searchString1.setPattern("error");
+        searchString1.setDisplayName("Error");
+        searchString1.setPrintMatchedString(false);
+
+        log.setSearchStrings(Lists.newArrayList(searchString, searchString1));
+
+        FilePointer filePointer = new FilePointer();
+        filePointer.setFilename(log.getLogDirectory() + File.separator + testFilename);
+        monitorContextConfiguration.setConfigYml("src/test/resources/conf/config.yaml");
+        when(mockFilePointerProcessor.getFilePointer(anyString(), anyString())).thenReturn(filePointer);
+
+        classUnderTest = new LogFileManager(mockFilePointerProcessor, log, monitorContextConfiguration);
+        LogMetrics logMetrics = classUnderTest.processLogMetrics();
+        Map<String, Metric> metrics = logMetrics.getMetrics();
+
+        assertEquals("3", metrics.get("active-dynamic-*|Search String|Debug|Occurrences").getMetricValue());
+        assertEquals("0", metrics.get("active-dynamic-*|Search String|Error|Occurrences").getMetricValue());
+
+        String filesize = getFileSize(log.getLogDirectory(), testFilename);
+        assertEquals(filesize, metrics.get("active-dynamic-*|File size (Bytes)").getMetricValue());
+
+        // simulate a file pointer update
+        filePointer.updateLastReadPosition(new Long(filesize));
+        when(mockFilePointerProcessor.getFilePointer(anyString(), anyString())).thenReturn(filePointer);
+
+        List<String> logsToAdd = Lists.newArrayList();
+        for (int i = 0; i < 100; i++) {
+            logsToAdd.add(new Date() + "	DEBUG	Statement " + i + "\n");
+        }
+        updateLogFile(testFilepath, logsToAdd);
+
+        // simulate new file created with different name
+        Thread.sleep(1000);
+        String dynamicLog2 = this.getClass().getClassLoader().getResource("dynamic-log-2.log").getPath();
+
+        testFilename = "active-dynamic-log-2.log";
+        testFilepath = String.format("%s%s%s", getTargetDir().getPath(), File.separator, testFilename);
+        copyFile(dynamicLog2, testFilepath);
+
+        // simulate another file created with different name
+        Thread.sleep(1000);
+        String dynamicLog3 = this.getClass().getClassLoader().getResource("dynamic-log-3.log").getPath();
+        testFilename = "active-dynamic-log-3.log";
+        testFilepath = String.format("%s%s%s", getTargetDir().getPath(), File.separator, testFilename);
+        copyFile(dynamicLog3, testFilepath);
+        logsToAdd.clear();
+
+        logMetrics = classUnderTest.processLogMetrics();
+        metrics = logMetrics.getMetrics();
+
+        assertNotSame(3, metrics.get("active-dynamic-*|Search String|Debug|Occurrences"));
+    }
+
+    @Test
+    public void testProcessorWhenEventsServiceIsEnabled() throws Exception {
+        Log log = new Log();
+        log.setDisplayName("TestLog");
+        log.setLogDirectory("src/test/resources/");
+        log.setLogName("test-log-1.log");
+
+        SearchString searchString = new SearchString();
+        searchString.setCaseSensitive(false);
+        searchString.setMatchExactString(true);
+        searchString.setPattern("debug");
+        searchString.setDisplayName("Debug");
+        searchString.setPrintMatchedString(false);
+
+        log.setSearchStrings(Lists.newArrayList(searchString));
+
+        FilePointer filePointer = new FilePointer();
+        filePointer.setFilename(log.getLogDirectory() + log.getLogName());
+        when(mockFilePointerProcessor.getFilePointer(anyString(), anyString())).thenReturn(filePointer);
+
+        MonitorContextConfiguration monitorContextConfiguration = new MonitorContextConfiguration("Log Monitor",
+                "Custom Metrics|Log Monitor|", Mockito.mock(File.class), Mockito.mock(AMonitorJob.class));
+        monitorContextConfiguration.setConfigYml("src/test/resources/conf/config-eventsService.yaml");
+
+        classUnderTest = new LogFileManager(mockFilePointerProcessor, log, monitorContextConfiguration);
+        LogMetrics logMetrics = classUnderTest.processLogMetrics();
+        Map<String, Metric> metrics = logMetrics.getMetrics();
+
+        assertEquals("13", metrics.get("TestLog|Search String|Debug|Occurrences").getMetricValue());
+
+        assertEquals(getFileSize(log.getLogDirectory(), log.getLogName()),
+                metrics.get("TestLog|File size (Bytes)").getMetricValue());
+    }
+
+    @Test
+    public void testLogEventGeneratorWithOffset() throws Exception {
+        EventsServiceDataManager eventsServiceDataManager = Mockito.mock(EventsServiceDataManager.class);
+        Mockito.when(eventsServiceDataManager.retrieveSchema(SCHEMA_NAME)).thenReturn("Hello world");
+        Log log = new Log();
+        log.setDisplayName("TestLog");
+        log.setLogDirectory("src/test/resources/");
+        log.setLogName("test-log-4-events-service.log");
+
+        SearchPattern searchPattern = new SearchPattern("Test Patterns", Pattern.compile("1"), false, false );
+
+        FilePointer filePointer = new FilePointer();
+        filePointer.setFilename(log.getLogDirectory() + log.getLogName());
+
+        OptimizedRandomAccessFile randomAccessFile = new OptimizedRandomAccessFile(new File("src/test/resources/test-log-4-events-service.log"), "r");
+
+        int offset = 5;
+
+        LogEventsProcessor classUnderTest = new LogEventsProcessor(eventsServiceDataManager, offset, log);
+        LogEvent logEvent = classUnderTest.processLogEvent(searchPattern, randomAccessFile, "");
+
+        Assert.assertEquals("1\n1\n2\n3\n4\n", logEvent.getLogMatch());
+    }
+    
+
+    
     private String getFileSize(String logDir, String logName) throws Exception {
         String fullPath = String.format("%s%s%s", logDir, File.separator, logName);
         RandomAccessFile file = new RandomAccessFile(fullPath, "r");
@@ -639,5 +694,4 @@ public class LogFileManagerTest {
         outputStreamWriter.write(sb.toString());
         outputStreamWriter.close();
     }
-    //endregion
 }